refactor: replace `displayAddress` with `chunked` method of a `PaymentCode` extension of `String`
